some timing rearrange
